Refactor ChunkedInferenceServiceResults to have chunksAsMatchedTextAndByteReference to allow each implementation to contain the code needed to convert the format for semantic text
