+++ x-pack/plugin/inference/src/main/java/org/elasticsearch/xpack/inference/mapper/SemanticTextField.java
@@ -8,7 +8,6 @@
 package org.elasticsearch.xpack.inference.mapper;
 
 import org.elasticsearch.ElasticsearchException;
-import org.elasticsearch.ElasticsearchStatusException;
 import org.elasticsearch.common.bytes.BytesReference;
 import org.elasticsearch.common.xcontent.XContentHelper;
 import org.elasticsearch.common.xcontent.support.XContentMapValues;
@@ -17,27 +16,21 @@ import org.elasticsearch.inference.ChunkedInferenceServiceResults;
 import org.elasticsearch.inference.Model;
 import org.elasticsearch.inference.SimilarityMeasure;
 import org.elasticsearch.inference.TaskType;
-import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.xcontent.ConstructingObjectParser;
 import org.elasticsearch.xcontent.DeprecationHandler;
 import org.elasticsearch.xcontent.NamedXContentRegistry;
 import org.elasticsearch.xcontent.ObjectParser;
 import org.elasticsearch.xcontent.ParseField;
-import org.elasticsearch.xcontent.ToXContent;
 import org.elasticsearch.xcontent.ToXContentObject;
-import org.elasticsearch.xcontent.XContent;
 import org.elasticsearch.xcontent.XContentBuilder;
 import org.elasticsearch.xcontent.XContentParser;
 import org.elasticsearch.xcontent.XContentParserConfiguration;
 import org.elasticsearch.xcontent.XContentType;
 import org.elasticsearch.xcontent.support.MapXContentParser;
-import org.elasticsearch.xpack.core.inference.results.ChunkedSparseEmbeddingResults;
-import org.elasticsearch.xpack.core.inference.results.ChunkedTextEmbeddingByteResults;
-import org.elasticsearch.xpack.core.inference.results.ChunkedTextEmbeddingFloatResults;
-import org.elasticsearch.xpack.core.ml.search.WeightedToken;
 
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -293,88 +286,13 @@ public record SemanticTextField(String fieldName, List<String> originalValues, I
     ) {
         List<Chunk> chunks = new ArrayList<>();
         for (var result : results) {
-            if (result instanceof ChunkedSparseEmbeddingResults textExpansionResults) {
-                for (var chunk : textExpansionResults.getChunkedResults()) {
-                    chunks.add(new Chunk(chunk.matchedText(), toBytesReference(contentType.xContent(), chunk.weightedTokens())));
-                }
-            } else if (result instanceof ChunkedTextEmbeddingFloatResults textEmbeddingFloatResults) {
-                for (var chunk : textEmbeddingFloatResults.chunks()) {
-                    chunks.add(
-                        new Chunk(
-                            chunk.matchedText(),
-                            toBytesReference(contentType.xContent(), chunk.embedding().getEmbedding().getFloats())
-                        )
-                    );
-                }
-            } else if (result instanceof ChunkedTextEmbeddingByteResults textEmbeddingByteResults) {
-                for (var chunk : textEmbeddingByteResults.getChunks()) {
-                    chunks.add(
-                        new Chunk(
-                            chunk.matchedText(),
-                            toBytesReference(contentType.xContent(), chunk.embedding().getEmbedding().getBytes())
-                        )
-                    );
-                }
-            } else {
-                throw new ElasticsearchStatusException(
-                    "Invalid inference results format for field [{}] with inference id [{}], got {}",
-                    RestStatus.BAD_REQUEST,
-                    field,
-                    inferenceId,
-                    result.getWriteableName()
-                );
+            for (Iterator<ChunkedInferenceServiceResults.Chunk> it = result.chunksAsMatchedTextAndByteReference(contentType.xContent()); it
+                .hasNext();) {
+                var chunkAsByteReference = it.next();
+                chunks.add(new Chunk(chunkAsByteReference.matchedText(), chunkAsByteReference.bytesReference()));
             }
         }
         return chunks;
-    }
-
-    /**
-     * Serialises the {@code value} array, according to the provided {@link XContent}, into a {@link BytesReference}.
-     */
-    private static BytesReference toBytesReference(XContent xContent, float[] value) {
-        try {
-            XContentBuilder b = XContentBuilder.builder(xContent);
-            b.startArray();
-            for (float v : value) {
-                b.value(v);
-            }
-            b.endArray();
-            return BytesReference.bytes(b);
-        } catch (IOException exc) {
-            throw new RuntimeException(exc);
-        }
-    }
-
-    private static BytesReference toBytesReference(XContent xContent, byte[] value) {
-        try {
-            XContentBuilder b = XContentBuilder.builder(xContent);
-            b.startArray();
-            for (byte v : value) {
-                b.value(v);
-            }
-            b.endArray();
-            return BytesReference.bytes(b);
-        } catch (IOException exc) {
-            throw new RuntimeException(exc);
-        }
-    }
-
-    /**
-     * Serialises the {@link WeightedToken} list, according to the provided {@link XContent},
-     * into a {@link BytesReference}.
-     */
-    private static BytesReference toBytesReference(XContent xContent, List<WeightedToken> tokens) {
-        try {
-            XContentBuilder b = XContentBuilder.builder(xContent);
-            b.startObject();
-            for (var weightedToken : tokens) {
-                weightedToken.toXContent(b, ToXContent.EMPTY_PARAMS);
-            }
-            b.endObject();
-            return BytesReference.bytes(b);
-        } catch (IOException exc) {
-            throw new RuntimeException(exc);
-        }
     }
 
 }
